+++ CHANGELOG.md
@@ -5,6 +5,27 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.3.1] - 2020-07-23
+
+### Fixed
+
+- Fix info.xml
+
+## [3.3.0] - 2020-07-23
+
+### New
+- Implement Activity app integration
+
+### Fixed
+- Add repair step for duplicate shared folders
+- Fix untagged search (on postgres)
+- UI: Fix rename Tag
+- Build: Don't ship source maps
+- Repair steps: Add debug output
+- Bookmarks: Additionally always sort by ID to make ordering stable
+- Fix deletion of bookmarks that are not in tree
+- Update dependencies
+
 ## [3.2.5] - 2020-07-17
 
 ### Fixed
@@ -271,6 +292,9 @@ Supported are NC 15 and 16, provided you are using PHP v7.1 and have gmp, intl a
 - FIX folder collapse css
 - FIX: Speed up findBookmarks SQL query
 
+
+[3.3.1]: https://github.com/nextcloud/bookmarks/compare/v3.2.5...v3.3.1
+[3.3.0]: https://github.com/nextcloud/bookmarks/compare/v3.2.5...v3.3.0
 [3.2.5]: https://github.com/nextcloud/bookmarks/compare/v3.2.4...v3.2.5
 [3.2.4]: https://github.com/nextcloud/bookmarks/compare/v3.2.2...v3.2.4
 [3.2.2]: https://github.com/nextcloud/bookmarks/compare/v3.2.1...v3.2.2